requirements.txt
# Core Deep Learning Frameworks (CUDA 11.8 Compatible)
torch==2.0.1+cu118
torchvision==0.15.2+cu118
torchaudio==2.0.2+cu118

# Mamba SSM and Related Dependencies
mamba-ssm==1.2.0.post1
triton==2.0.0

# Diffusion Models
diffusers==0.29.2
transformers==4.40.2
accelerate==0.20.3

# HuggingFace Ecosystem
huggingface-hub==0.34.1
tokenizers==0.19.1
safetensors>=0.3.1

# Audio Processing
librosa>=0.9.0
scipy>=1.9.0

# Video Processing (requires FFmpeg and libx264)
moviepy>=1.0.3
imageio>=2.31.1
imageio-ffmpeg>=0.4.9

# Image Processing and Computer Vision
opencv-python>=4.5.0
Pillow>=8.0.0
torchvision>=0.15.0
mediapipe>=0.10.0

# Scientific Computing
numpy==1.24.3
einops>=0.6.0

# Configuration Management
omegaconf>=2.2.0
pyyaml>=5.1

# Progress Bars and Utilities
tqdm>=4.64.0
packaging>=20.9

# File Processing
fsspec>=2023.5.0
filelock>=3.8.0

# Regular Expressions
regex!=2019.12.17

# HTTP Requests
requests>=2.28.0
urllib3<1.27,>=1.21.1
charset-normalizer<3,>=2
idna<4,>=2.5
certifi>=2017.4.17

# Type Checking
typing-extensions>=3.7.4.3

# Other Dependencies
importlib-metadata>=4.0.0
zipp>=3.20
psutil>=5.8.0
ninja>=1.10.0
cmake>=3.18.0
lit>=15.0.0
MarkupSafe>=2.0
mpmath<1.4,>=1.1.0
networkx>=2.8.0
jinja2>=3.0.0
sympy>=1.10.0

# HF-XET (Optional, for large file downloads)
hf-xet<2.0.0,>=1.1.3